### 2020-07-01 #2416 @whedon
[PRE REVIEW]: Chromatiblock: scalable whole-genome visualization of structural differences in prokaryotes
**Submitting author:** @mjsull (<a href="http://orcid.org/0000-0001-9076-6963">Mitchell</a>)
**Repository:** <a href="https://github.com/mjsull/chromatiblock/" target ="_blank">https://github.com/mjsull/chromatiblock/</a>
**Version:** v0.4.2
