### 2020-07-06 #2416 @will-rowe
Hi @abotzki - would you also be able to review this submission for JOSS?
