### 2020-07-07 #2416 @will-rowe
Thanks @happykhan - and thank you very much for the suggestions. @leomrtns, @lskatz and @telatin - would any of you be able to review this (even better if two of you could). I think this looks like a really useful tool for our field and it would be great to get your takes on it.
