### 2020-07-08 #2416 @will-rowe
Hey @rpetit3 and @coughls. Would either of you be able to review this submission for JOSS please?
