### 2020-07-08 #2451 @whedon
[REVIEW]: Chromatiblock: scalable whole-genome visualization of structural differences in prokaryotes
**Submitting author:** @mjsull (<a href="http://orcid.org/0000-0001-9076-6963">Mitchell</a>)
**Repository:** <a href="https://github.com/mjsull/chromatiblock/" target ="_blank">https://github.com/mjsull/chromatiblock/</a>
**Version:** v1.0.0
