### 2020-07-09 mjsull/chromatiblock#10 @telatin
JOSS Review
:link: https://github.com/openjournals/joss-reviews/issues/2451

I'll put in this issue my comments. So far I really enjoyed trying this tool that looks really useful.

## Paper

* _C. difficile_ is not italicized

## Documentation and repository structure

The section "**Demo**" should be expanded (as it could be the very first attempt of using the software by a user) adding:

* A short description of the dataset
* A short description of the task to accomplish (_e.g._ use of custom categories colours)
* A link to the expected output
(_H. pylori_ output is linked in the paragraph above and can be easily missed)

A section describing the produced **output**:

* The `-o` switch can be described better, it's not immediately clear that can be used to select a different output format by changing the extension.
* A concise guide to interpreting the image produced
* An explanation of the intermediate files produced (or it could be an idea to delete the intermediate folder by default,
unless `--keep-intermediates` or similar is specified. In this way it would not be so necessary to describe the files)

Minor issues:

* License file contains placeholders:
`<program>  Copyright (C) <year>  <name of author>`
* An explicit "CONTRIBUTING.md" could be helpful,
as it's common practice to have one
* Readability of README.md  can be improved. _Examples:_
  * Version and License are rendered as subtitles,
  * Program flags are rendered as titles and they look very big on a browser.
I assume they are automatically generated from the documentation
so they can remain as they are if it's unpractical to change this.
  * Use of lists could improve readability (_e.g._ requirements, parameters,...)

## Usability

_Summary_: the program works as expected and is simple to run.

### Dependency check

The program if invoked with `-h` provides the expected documentation.
If invoked without parameters will print "_No input files found use -f or -d_", and I would add "_Type `--help` for usage_".

Downloading the repository and trying to run the program without installing the dependencies results in:

```bash
/bin/sh: Sibelia: command not found
Traceback (most recent call last):
  File "./chromatiblock", line 1548, in <module>
    block_dict = get_blocks(args.working_directory, name_dict, seq_dict, gene_list, cats)
  File "./chromatiblock", line 690, in get_blocks
    with open(sibel_dir + '/blocks_coords.txt') as f:
FileNotFoundError: [Errno 2] No such file or directory: 'simple_wd/blocks_coords.txt'
```

For beginner users, this might not be an easy to understand message, and I would *suggest* to consider adding a "dependency check".

### Output

Here I found some issues:

1. The produced image should have the labels for each genome [edit: see #9].
2. Trying to produce an _svg_ output failed for me (conda installation, command issued `chromatiblock -d simple_demo -w simple_wd -o H_pylori.pdf -l simple_demo/order_list.txt`, output [here](https://gist.github.com/telatin/68f0df49a803aeea89d10244a6b65aa3)). I couldn't open the file with Inkscape.
3. Trying to produce a _pdf_ output failed (conda installation):
```
Traceback (most recent call last):
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/xml/etree/ElementTree.py", line 1693, in feed
    self.parser.Parse(data, 0)
xml.parsers.expat.ExpatError: not well-formed (invalid token): line 3, column 9

Traceback (most recent call last):
  File "./chromatiblock", line 1555, in <module>
    draw_blocks(core_array, noncore_pos, core_size, args.out, args.genome_height, args.gap, legend_size, args.working_directory, args.ppi, args.categorise != None, args.svg_pan_zoom_location, args.hue_start, args.hue_end)
  File "./chromatiblock", line 1351, in draw_blocks
    cairosvg.svg2pdf(url=os.path.join(working_dir, "temp.svg"), write_to=out_file, dpi=ppi)
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/site-packages/cairosvg/__init__.py", line 76, in svg2pdf
    return surface.PDFSurface.convert(
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/site-packages/cairosvg/surface.py", line 142, in convert
    tree = Tree(
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/site-packages/cairosvg/parser.py", line 405, in __init__
    tree = ElementTree.fromstring(
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/site-packages/defusedxml/common.py", line 131, in fromstring
    parser.feed(text)
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/xml/etree/ElementTree.py", line 1695, in feed
    self._raiseerror(v)
  File "/Users/telatina/miniconda3/envs/chromatiblock/lib/python3.8/xml/etree/ElementTree.py", line 1602, in _raiseerror
    raise err
xml.etree.ElementTree.ParseError: not well-formed (invalid token): line 3, column
```

## Code

The code lacks comments, and this makes it hard to contribute to it. This should be improved before publication.

_Comment/Question_: Both SVG and HTML are hardcoded in the script. While I understand that this makes more portable the script and removes the problem of missing external files, I wonder if using specialized libraries would make the program more future-proof.

## Tests

Since there are no automatic tests, it would be useful to add a bash script that automatically:
- Downloads the simple example
- Produces all possible output formats
- Checks the produced files
- Summarizes which succeeded/failed
