### 2020-08-06 #2451 @mjsull
Hi Will,

Sorry about missing your earlier email. Definitely very helpful, working
through the comments now. Will finish over the weekend.

Best,

Mitch
