### 2020-09-08 1c74881
Addressing the issues raised by JOSS reviewers - fixed formatting issues in paper.md - Made README clearer. - more comments in the main script - added flag -t to draw labels for each fasta - added flag --force to forec overide - added flag --keep to keep intermediate files (now automatically deleted - added flag --output_format so user needs to specify format - added flag --extensions to define which files in a directory will be used
