+++ README.md
@@ -1,12 +1,11 @@
 # chromatiblock
 
 ## Scalable, whole-genome visualisation of structural changes in prokaryotes
+![version](https://img.shields.io/badge/version-0.4.2-blue) ![license](https://img.shields.io/badge/license-GPLv3-green)
 
-### Version: 0.4.2
+### Installation 
 
-### License: GPLv3
-
-### Installation via conda:
+#### Conda:
 
 To create a fresh environment for chromatiblock to run in do:
 
@@ -20,22 +19,19 @@ Then in future to run chromatiblock you can reactivate this environemtn using ``
 
 #### Direct download:
 
-Alternatively you can download and run the script from [here](https://github.com/mjsull/chromatiblock/releases/download/v0.4.1/chromatiblock).
+Alternatively you can download and run the script from [here](https://github.com/mjsull/chromatiblock/archive/v0.4.2.zip).
 
 #### requirements:
 (these will be installed automatically if chromatiblock is installed with conda)
 
-**Python >= 3.6.0**
+- **Python >= 3.6.0**
 
-While chromatiblock will run fine without these programs, they are needed for certain tasks.
+While chromatiblock will run fine without the following programs, they are needed for certain tasks.
 
-Sibelia - for autmoatic generation of colinear blocks
-
-cairosvg - for creating PNG and PDF images (svg and html supported natively) 
-
-ncbi-blast+ - for automatic annotation of genes
-
-svg-pan-zoom.min.js - by default the html output uses an online javascript library for zoom functionality,
+- **Sibelia** - for automatic generation of colinear blocks
+- **cairosvg** - for creating PNG and PDF images (svg and html supported natively) 
+- **ncbi-blast+** - for automatic annotation of genes
+- **svg-pan-zoom.min.js** - by default the html output uses an online javascript library for zoom functionality,
 the -pz flag can be used to point to an offline location of svg-pan-zoom.
 
 ### Example usage: 
@@ -52,32 +48,46 @@ or
 Will run chromatiblock on all fasta or genbank files in `/path/to/fasta_directory/` and create an interactive webpage of the output. Intermediate files will be created
 in `cb_working_dir`.
 
-# Example output:
+### Example output:
 
 [![chromatiblock](https://mjsull.github.io/chromatiblock/images/chromatiblock_main.gif)](https://mjsull.github.io/chromatiblock/C_difficile.html)
 
-### An interactive HTML example can explored by clicking on the above figure.
+#### An interactive HTML example can explored by clicking on the above figure.
 
 A tooltip will show block # and location when the mouse is hovered over a block. Other blocks of the same type will also be highlighted. Clicking on a block will show additional information, including the underlying sequence.
 
 ### Figure description
 
-An example of a global alignment of 28 complete C. difficile ge-nomes is shown in the above figure. 
-Multi-locus sequence type (MLST) of each isolate is indicated on the left.
+An example of a global alignment of 28 complete *C. difficile* genomes is shown in the above figure. 
 
-**Top)** Global alignment view. Core blocks across genomes (rows) are visualized as verti-cally aligned solid rectangles that are colored according to their position in the genome. 
-Non-core blocks are visualized as patterned rectangles, with each block represented by a unique combination of pattern and color. Finally, sequences unique to a single genome are depicted as solid black lines.  
+**TOP)** The global alignment view shows the arrangement of core blocks (i.e. syntenic regions found
+once in all genomes) in the alignment and how non-core blocks (i.e syntenic regions found in 2 or
[…]
