### 2020-09-08 7c34745
updated the README with new flags and version
