### 2020-09-09 0568392
small fixes to README.md
added a test script

small changes to the epilogue of chromatiblock argparse
