### 2020-09-09 mjsull/chromatiblock#10 @mjsull
Hi Dr. Telatin,

Thank-you for your review, your comments were very helpful.

I hope I have addressed them all below.

## Paper
>C. difficile is not italicized

This has been fixed in the revised manuscript.

## Documentation and repository structure
>The section "Demo" should be expanded (as it could be the very first attempt of using the software by a user) adding:
A short description of the dataset

We have now added a short description of the dataset in the ‘Demo’ section.

> A short description of the task to accomplish (e.g. use of custom categories colours)

This description has now been added.

> A link to the expected output
 (H. pylori output is linked in the paragraph above and can be easily missed)

This has been moved to the Example section

> The -o switch can be described better, it's not immediately clear that can be used to select a different output format by changing the extension.

We have modified the behavior of this switch so the user has to specify format output

> A concise guide to interpreting the image produced

More information has been added to the figure included in the README

> An explanation of the intermediate files produced (or it could be an idea to delete the intermediate folder by default,
 unless --keep-intermediates or similar is specified. In this way it would not be so necessary to describe the files)

Intermediate files are now deleted unless the --keep flag is used. A short description of the intermediate files has also been added to the  README

## Minor issues:
>License file contains placeholders:
 <program> Copyright (C) <year> <name of author>

These placeholders have now been removed from the license file as per GPL instructions

>An explicit "CONTRIBUTING.md" could be helpful,
 as it's common practice to have one

The contributing information has now been added

>Readability of README.md can be improved. Examples:
Version and License are rendered as subtitles, they can be written as simple text on a single line (or badges like )
Program flags are rendered as titles and they look very big on a browser.
 I assume they are automatically generated from the documentation
 so they can remain as they are if it's unpractical to change this.
Use of lists could improve readability (e.g. requirements, parameters,...)

Additional work has been done to improve the layout and formatting of the README.md file

## Usability
> Summary: the program works as expected and is simple to run.
Dependency check
The program if invoked with -h provides the expected documentation.
If invoked without parameters will print "No input files found use -f or -d", and I would add "Type --help for usage".

An explanation of the --help flag has been added to the output to make it more easy to discover.

>Downloading the repository and trying to run the program without installing the dependencies results in:
/bin/sh: Sibelia: command not found
Traceback (most recent call last):
  File "./chromatiblock", line 1548, in <module>
    block_dict = get_blocks(args.working_directory, name_dict, seq_dict, gene_list, cats)
  File "./chromatiblock", line 690, in get_blocks
    with open(sibel_dir + '/blocks_coords.txt') as f:
FileNotFoundError: [Errno 2] No such file or directory: 'simple_wd/blocks_coords.txt'
For beginner users, this might not be an easy to understand message, and I would suggest to consider adding a "dependency check".

This dependency check has now been added.

## Output
>Here I found some issues:
The produced image should have the labels for each genome [edit: see #9].

Contig names can now be seen by hovering the mouse in html mode. An extra flag to add labels generated from the FASTA file names has also been implemented.

>Trying to produce an svg output failed for me (conda installation, command issued chromatiblock -d simple_demo -w simple_wd -o H_pylori.pdf -l simple_demo/order_list.txt, output here). I couldn't open the file with Inkscape.
Trying to produce a pdf output failed (conda installation):

This was caused by a problem with the svg format, it has now been fixed.

Code
>The code lacks comments, and this makes it hard to contribute to it. This should be improved before publication.

Additional comments have been added throughout the code to facilitate external contributions.

>Comment/Question: Both SVG and HTML are hardcoded in the script. While I understand that this makes more portable the script and removes the problem of missing external files, I wonder if using specialized libraries would make the program more future-proof.

The goal was to enable the user to be able to interact with the svg in the generated webpage (tooltips/html links/highlighting). Several javascript solutions were tried, but all came with serious performance issues. We arrived at a solution involving some custom CSS, however this required that tags/groups be included in the objects in the svg file. To our knowledge, this feature is not available in open source svg libraries.

## Tests
>Since there are no automatic tests, it would be useful to add a bash script that automatically:
Downloads the simple example
Produces all possible output formats
Checks the produced files
Summarizes which succeeded/failed

We have now added a basic test case that implements the steps outlined by this reviewer.
