### 2020-09-09 #2451 @mjsull
Hi @will-rowe

I think I've addressed everything. Please find my comments for each reviewer in their issue threads.

https://github.com/mjsull/chromatiblock/issues/10
https://github.com/mjsull/chromatiblock/issues/12

Best,

Mitch
