### 2020-09-09 #2451 @rpetit3
@mjsull everything looks great! I have merged your latest release into Bioconda (https://github.com/bioconda/bioconda-recipes/pull/24287). Once its synced to Anaconda's CDN, I'll give it a test run and report back (hopefully today).
