### 2020-09-09 mjsull/chromatiblock#12 @rpetit3
Hi @mjsull

This is great! I tested out 0.5.1 and everything finished as expected. The PDF, PNG and SVG all look good. (Really like the `-of all` option!).

Thank you very much for your hard work!
Robert
