### 2020-09-09 #2451 @rpetit3
Hi @will-rowe

I have tested @mjsull's latest release of Chromatiblock (v0.5.1) and everything looks good. The only major hold up on my end was the SVG, PNG, PDF output, which is now fixed in the latest version. I think @mjsull really went above and beyond responding to the rest of my comments.

I am happy to recommend this manuscript for publication!

Thank you for your time!
Robert
