### 2020-09-10 #2451 @will-rowe
Hi @mjsull - this is looking really nice. Just playing with the software now and am having no difficulties. The paper is also reading nicely. The only thing needing changing from my point of view is a fix for the Sibelia reference.

Thank you again @telatin and @rpetit3 - your reviews were excellent and this submission is looking great. Well done @mjsull with the responses.

@telatin has offered to take a look at your responses by end of the week - so we will wait on him to have a final look. Once has gives his approval - I'm happy for this to proceed to publication.
