### 2020-09-11 mjsull/chromatiblock#10 @telatin
Hello @mjsull, thanks for releasing an improved version of your tool. I really appreciated the `--add_fasta_labels` option.
Fine by me, i'll complete the review in JOSS now.

As a suggestion, if you intend to keep updating the tool in the future, I would expand the test script to become a "unit test" like:
```
dependencies: OK
printing version: OK

generate SVG from 4 fasta: OK
generate HTML from 4 fasta: OK
generate SVG using list: OK
etc
```
this way, when adding features you'll spot if something you didnt intend to alter has been impacted. But I understand this might be an overkill so its fine by me as it is (minor hint: I already had downloaded the file and this made the script blocking execution as the "unzip" command prompted for overwrite... some checks can still be beneficial)
