+++ paper/paper.bib
@@ -114,7 +114,8 @@ https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3031037/},
    publisher = {Springer Berlin Heidelberg},
    pages = {215-229},
    ISBN = {978-3-642-40453-5},
-   type = {Conference Proceedings}
+   type = {Conference Proceedings},
+   DOI = {10.1007/978-3-642-40453-5_17}
 }
 
 @article{RN1,
