### 2020-09-17 #2451 @mjsull
Hi Will,

chromaitblock DOI has been added.

v1.0.0 released, the DOI: 10.5281/zenodo.4034604
