### 2020-09-18 #2451 @kyleniemeyer
Hi @mjsull, I'm the EIC on duty this week, and doing some final checks before publishing. Could you address these issues?

- Please edit the metadata on the Zenodo archive so that the author names match what is in the paper. (By default Zenodo tends to just grab the GitHub usernames of contributors).
- Your paper is missing an explicit Statement of Need section, [required for all JOSS papers](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which should briefly illustrate the research purpose of the software. You can probably use some material given in the Introduction
- Also, can you break up the Introduction paragraph somewhat? Though this might be satisfied by fixing the previous issue
- please put commas after "i.e." throughout (I noticed mostly in Implementation section)
