### 2020-09-23 2cded04
final fixes to paper
