+++ paper/paper.md
@@ -33,27 +33,33 @@ Microsoft Windows from https://github.com/mjsull/chromatiblock/
 Visualizing structural variation between complete prokaryotic genomes is
 important for identifying the genetic basis of strain differences. This is
 generally accomplished by displaying the results of serial pairwise comparisons
-or multiple alignments in linear or circular layouts. Serial pairwise
-comparisons can be created using tools such as Easyfig [@RN1] or  GenoplotR [@RN2]
+or multiple alignments in linear or circular layouts. Serial pairwise comparisons can be created using tools such as Easyfig [@RN1] or  GenoplotR [@RN2]
 that display linear pairwise comparisons between two or more genomes. However, 
 genomic loss, gain and structural variation can only be directly inferred for
-genomes adjacent to each other. Multiple alignment visualization tools such as
+genomes adjacent to each other. 
+
+Multiple alignment visualization tools such as
 Mauve [@RN3] and GenomeRing [@RN4] solve this issue by representing syntenic
 regions as blocks and using lines to connect blocks across genomes or to
 indicate block order, respectively. In large figures this can result in
-crisscrossing lines that are often difficult to interpret. Alternatively, ring
-plots, such as those created by the BLAST ring image generator (BRIG) [@RN5] or
+crisscrossing lines that are often difficult to interpret.
+Alternatively, ring plots, such as those created by the BLAST ring image generator (BRIG) [@RN5] or
 the CGView Comparison Tool (CCT) [@RN6] use a series of concentric circles to
 display the presence or absence of genomic regions across multiple genomes.
 These regions are ordered according to a reference, and as such they convey
-no information about their arrangement in each non-reference genome.
-Alternatively, Circos [@RN7] plots show genomes around the outside edge of a
-circle and represents regions of similarity as arcs, but this approach scales
-poorly as the number of arcs increases exponentially with each genome.
+no information about their arrangement in each non-reference genome. 
 Representing many genomes as circles can also result in large size differences
-between inner and outer rings, further complicating interpretation. Here we
-present Chomatiblock, an application for visualizing syntenic blocks in
-multiple genome alignments. Chromatiblock was designed to create a linear
+between inner and outer rings, further complicating interpretation. Circos [@RN7] plots show genomes around the outside edge of a
+circle and represents regions of similarity as arcs, but this approach scales
+poorly as the number of arcs increases exponentially with each genome. 
+Here we present Chomatiblock, an application for visualizing syntenic blocks in
+multiple genome alignments.
+
+# Statement of Need
+Current methods of visualizing multiple genomes either don't scale well, or only 
+capture limited information about structural variation. With complete genomes becoming
+more prevalent, there is an need for a tool that can do both well.
+Chromatiblock was designed to create a linear
 visual representation of structural variation, including the presence and
 absence of genomic regions in an easy-to-comprehend and scalable manner, adding
 to the visualization options available for alignments of large numbers of
@@ -68,16 +74,15 @@ syntenic blocks. The results can also be viewed as an interactive webpage that
 allows the user to zoom, pan and highlight shared regions across genomes.
 
 Chromatiblock takes an extended multi-fasta alignment (MAF) file as input, which
-can be generated by a variety of multi-genome alignment programs (Angiuoli and
-Salzberg, 2011; Minkin and Medvedev, 2019). Alternatively, when provided with
+can be generated by a variety of multi-genome alignment programs [@RN9;@RN10]. Alternatively, when provided with
 FASTA-formatted files for a set of genomes of interest, Chromatiblock can run
-Sibelia (Minkin, et al., 2013) to automatically generate the syntenic blocks
+Sibelia [@RN8] to automatically generate the syntenic blocks
[…]
