### 2020-09-23 #2451 @kyleniemeyer
Many thanks to @telatin, @rpetit3, and @rpetit3 for reviewing this, and @will-rowe for editing it.

I see that the paper appears at https://joss.theoj.org/papers/10.21105/joss.02451, but the DOI does not yet resolve, so I'll keep this issue open until it does.
