### 2020-09-24 #2451 @kyleniemeyer
@rpetit3 oh, that's strange, I totally missed that!

@openjournals/dev can you fix this? @rpetit3 shows up twice as reviewer on the paper (likely due to showing up in the review issue twice—perhaps due to being added back at some point?)
