### 2020-09-24 #2451 @will-rowe
Hi @kyleniemeyer and @openjournals/dev - this may have been my fault. For some reason, @rpetit3 was added to review as usual but had no access writes to the issue. We removed and added again and it fixed it, but not sure why he is listed twice! Sorry.
