### 2020-09-24 #2451 @arfon
> @openjournals/dev can you fix this? @rpetit3 shows up twice as reviewer on the paper (likely due to showing up in the review issue twice—perhaps due to being added back at some point?)

This is now fixed but may take some time to update on the JOSS site due to caching.
